One blank view. Loads up gps. Logs coords to log.
JTViewController gets initialized and viewDidLoad gets called when the
view is loaded. Thats when we setup gps, which prompts the user. We set
ourselves as location services delegate so delegate methods will be
called on our instance when the location is updated. Chiefly
locationManager:didUpdateLocations:. This method is defined on our
controller to log to the console.